Revise codes
