+++ src/document/crdt/element.ts
@@ -15,7 +15,7 @@
  */
 
 import { TimeTicket } from '@yorkie-js-sdk/src/document/time/ticket';
-import type * as Devtools from '@yorkie-js-sdk/src/types/devtools';
+import type * as Devtools from '@yorkie-js-sdk/src/devtools/types';
 
 /**
  * `CRDTElement` represents an element that has `TimeTicket`s.
