+++ src/document/crdt/primitive.ts
@@ -19,7 +19,7 @@ import { Code, YorkieError } from '@yorkie-js-sdk/src/util/error';
 import { TimeTicket } from '@yorkie-js-sdk/src/document/time/ticket';
 import { CRDTElement } from '@yorkie-js-sdk/src/document/crdt/element';
 import { escapeString } from '@yorkie-js-sdk/src/document/json/strings';
-import type * as Devtools from '@yorkie-js-sdk/src/types/devtools';
+import type * as Devtools from '@yorkie-js-sdk/src/devtools/types';
 
 export enum PrimitiveType {
   Null,
