+++ src/document/document.ts
@@ -69,7 +69,7 @@ import {
 } from '@yorkie-js-sdk/src/document/presence/presence';
 import { History, HistoryOperation } from '@yorkie-js-sdk/src/document/history';
 import { setupDevtools } from '@yorkie-js-sdk/src/devtools';
-import * as Devtools from '@yorkie-js-sdk/src/types/devtools';
+import * as Devtools from '@yorkie-js-sdk/src/devtools/types';
 
 /**
  * `DocumentOptions` are the options to create a new document.
